Update patch set 4
Patch Set 4:

Issues 1 and 2 fixed.

Issue 3 is on master and is caused by the hacky way in which I invoke search so that the keyboard on Android pops out. In other words, I had to choose, either the keyboard is opened when search is opened, or the forward button (which probably nobody would use in this case on mobile) works. I consulted this with Kaity, and she preferred the keyboard.

Patch-set: 4
Reviewer: JGonera <542@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0