Use Special:MobileLanguages in beta/alpha
Don't load langlinks when initially loading the
page; instead, wait until button is pressed.

If JavaScript is available, load the langlinks via
the api and then show a LanguageOverlay; if not, just
link to Special:MobileLanguages.

Also remove now-unneeded parseList() function and associated
tests.

This is all done in beta (when it goes stable, there are
redundant functions that should be removed and have been
marked as such).

Bug: 40678
Change-Id: I7d5f2fa9c6d30fe2f2ca2d194f6ecbe72b02c88b