Update patch set 7
Patch Set 7:

It masks a bigger problem that we will need to find a nice way to handle - the images are relative to where the styles are generated. In the case of KSS documentation a css file is generated from mobile.less that is put in public/style.less - all the images will point as if style.less lives in the same folder as mobile.less which it doesn't.

It's hacky but I'm not sure how best to resolve this :(

Patch-set: 7
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0