Update patch set 2
Patch Set 2:

(3 comments)

In terms of your other comments - the problem is the design [1] in the story has various holes - for instance it doesn't tell me what to do with the last modified bar. With the absence of anyone in the office I've been having trouble trying to work out what to do so have had to make some assumptions.

"This adds left and right margin to the whole page, those margins should only affect the bottom border of the header." I guess they could be margins on the last modified bar rather than the content wrapper but that looked a little odd to me.

As stated in the commit message I didn't expect this to be perfect first time round and I had to stop as I was getting super confused by the overlay css rules. Feel free to pick it up and bring the patchset further along but I wanted to get an early version out of there as I felt this was progress in the right direction.

[1] https://www.dropbox.com/s/xfaxym40h9dv0ov/updated-header-03.png

Patch-set: 2
Label: Verified=0