Update patch set 2
Patch Set 2:

(1 comment)

"for instance it doesn't tell me what to do with the last modified bar"

Fair point. I would assume though that the only thing that should have left/right margin is the horizontal line just below the header, nothing else. But let's clarify with Kaity. I've sent an e-mail.

"I guess they could be margins on the last modified bar rather than the content wrapper but that looked a little odd to me."

I wasn't talking about adding margins to the last modified bar, I was talking about _not_ adding them there. Presently there is additional left/right white margin on the whole page.

"As stated in the commit message I didn't expect this to be perfect first time round and I had to stop as I was getting super confused by the overlay css rules. Feel free to pick it up and bring the patchset further along but I wanted to get an early version out of there as I felt this was progress in the right direction."

Considering that we want to push it to stable in the following iteration, I think it should be as close to perfect as possible. Unless we want to postpone it further. I doubt I'll have time to work on this today, but I can have a look on Monday.

Patch-set: 2