Update patch set 4
Patch Set 4:

Made a second pass. Had to guess on the header styling for special pages again as was unable to get hold of Kaity on IRC.

"Considering that we want to push it to stable in the following iteration, I think it should be as close to perfect as possible. Unless we want to postpone it further. I doubt I'll have time to work on this today, but I can have a look on Monday."

Yes agreed. I wasn't saying we wouldn't want it perfect when pushing to stable, I was just keen to get designers looking at some kind of version on a labs instance that they can tweak (proving difficult to do design right now with most people remote) and avoid rebases.

That said this probably isn't quite ready but it's close. Feel free to take a stab at this on Monday.

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0