Update patch set 8
Patch Set 8:

Juliusz this was a horrible coding experience but I think I've got the required behaviour now. My brain is now fried for the weekend.

Feel free to pick this up on Monday and fix anything that might have been broken in this refactor. I need to take a step away from this code. :)
Happy weekend!

Patch-set: 8