Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

Seems to cause a regression in stable - when I click edit on a page in stable and make changes, when the page reloads and I click the language button there are no languages in the list. Seems fine in beta.

The code in Page.js needs a quick rethink since getPageLanguages now resolves a deferred objects with both languages and variants.

Patch-set: 3
Label: Code-Review=-1