Update patch set 3
Patch Set 3:

In general it would be nice if we could disable section toggling altogether.

I'm also ok with the changes made in here (although I'm not into MobileFrontend enough to rate them properly). I'm not a fan of not having this per default (especially as disabled people probably wont even be able to switch into another mode due to other bugs(?))

Patch-set: 3