Story 1516: New A/B test for Sign-up Edit Guider
If the user signs up from the left nav and they belong to the test
group, load an edit guider when they return to the article they
came from.

Also fixing the fact that edit guiders are currently broken due to the
edit button not being activated in time. Now mobile.editor is
explicitly loaded first (as a dependency).

Change-Id: Ibce39dd63bd5200a4d68b34efcf3a3a2abd09d28