Update patch set 1
Patch Set 1: Code-Review-1

This works great thanks!
That said we need to be more cautious and limit this fix to alpha. Currently the CSS loads in beta and stable as well and may have side effects.

Update the rules in mainmenuAnimation.less (which is only loaded in alpha) and leave mainMenu.less unchanged and I'll merge this.

Patch-set: 1
Label: Code-Review=-1