Update patch set 6
Patch Set 6: Code-Review-1

Please break into two patchsets as suggested above.
I think the JavaScript code is useful, but I'm a bit concerned that focusing too much on the server side is probably not worth our while since this is a lot of code for a limited use case. I'm happy to sign off the Code In task and bug with just the JavaScript use case addressed.

Patch-set: 6
Label: Code-Review=-1