Update patch set 4
Patch Set 4:

The only difference between the existing step and your step is that your new step has a hard-coded file name while the existing step takes the file name as an argument from the .feature file.  

It doesn't matter if you change the wording of the step, but it is much better to re-use existing code, and it is much better not to hard-code things like file names.  

Your step:
----
 require 'tempfile'
  path = "#{Dir.tmpdir}/image.png"

  require 'chunky_png'
  ChunkyPNG::Image.new(Random.new.rand(255), Random.new.rand(255), Random.new.rand(255)).save path

  on(LeadPhotoPage).select_file_element.send_keys(path)
----

and the step that already exists: 

----
require 'tempfile'
  path = "#{Dir.tmpdir}/#{file_name}"

  require 'chunky_png'
  ChunkyPNG::Image.new(Random.new.rand(255), Random.new.rand(255), Random.new.rand(255)).save path

  on(UploadsPage).select_file_element.send_keys(path)
----

Patch-set: 4