Update patch set 10
Patch Set 10:

The workflow of having to go to BetaFeatures to enable the skin, and then having to go to another section of your preferences to actually turn it on seems weird to me. Can we just have the user turn it on via the BF panel directly?

Is the image an actual screenshot from Wikipedia? If so attribution should be provided and copyright needs to be indicated somewhere.

Patch-set: 10
Reviewer: Legoktm <458@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0