Update patch set 5
Patch Set 5: Code-Review-1

(5 comments)

I may be wrong but there seems to be a lot of minor changes to stable. Can we be more conservative?
Also is there any way we can split this commit up to make it more readable?
e.g.
1) Create content.less
2) Create beta/content.less etc..

Also I dislike the New suffix - should probably talk about this as a team and find a better convention (don't worry about that here for the sake of this patch though).

Also for the record I love it all apart from the fact I hugely dislike the centered text in the page on large screens  (I thought it was a bug) although after talking to Kaity I understand this by design.. :)

Patch-set: 5
Label: Code-Review=-1