Update patch set 1
Patch Set 1:

So the main problem with this is the code doesn't actually work - when I set $wgMFAnonymousEditing to true I get no warning message.

The other problem is that currently the overlays are being completely redesigned which will mean this code will soon not exist. You'll want to apply any changes to EditorOverlayNew

That said the editor was designed to be full screen without scrolling so I seriously worry about porting a warning message in a similar way to how the editor works on desktop - especially in languages where the translated string is long. I suspect with the fixed headers this will leave a tiny textarea for editors which is a serious UX problem. I think potentially an intermediate screen may be better in this context but I'm not a designer so I'm not sure. This would be a good thing to discuss on mailing lists first and get some guidance on.

To be clear, I'm keen to see anonymous editing happen and I'm glad you are poking at this, I just want to ensure we give it the user experience it deserves.

Patch-set: 1