Removing CSS reset for border-spacing. Making table headers bold.
O is a bad default for border-spacing since it causes data in
adjacent cells to run into each other.

Table headers should be bold unless there's a good reason for them
not to be. Bold is the browser default and how they are displayed
on desktop. If they are not displayed in bold it is not always
clear that they are headers rather than regular cells.

Change-Id: I2df83a6038e5ede5e8687520dee3dad7efef7d21