Update patch set 5
Patch Set 5: Code-Review-1

As a new user when I go to a page which needs a lead photo clicking on the upload button should open the tutorial. This happens in beta but not stable.

This is described in https://gerrit.wikimedia.org/r/105106 which would have caught this problem - can we get that patchset merged and then adjusted along with this patchset to ensure we don't break anything?

Patch-set: 5
Label: Code-Review=-1