Update patch set 10
Patch Set 10: Code-Review-1

It would be good for this to be packaged with a new feature test to show how it will be used. I worry about merging code before this has happened as it's possible this code will not do what we want it to do or will be forgotten about and simply confuse newcomers to the repository.

Patch-set: 10
Label: Code-Review=-1