Update patch set 9
Patch Set 9:

Hi Jared
If you open up developer tools and watch the network tab, you'll be able to see how this queries the API and how this breaks in the stable mode of the site. Alternatively you could rebase your patchset off https://gerrit.wikimedia.org/r/#/c/106659/ which will simplify the codebase and make this less of a concern.

Patch-set: 9