QA: Add a short pause for notification to disappear
Note: this is a deliberate use of sleep() as opposed to a
indeterminate wait.  If after clicking BACK or CLOSE the
notification overlay takes longer than one second to disappear,
we want to know that, as opposed to a variable wait of up to
several seconds.

This is causing the failure noted at
https://wmf.ci.cloudbees.com/job/MobileFrontend-en.m.wikipedia.beta.wmflabs.org-linux-chrome/229/

Change-Id: Icf57cbebd13ba83bd4c50ebe2f185204de9ef24e