Update patch set 1
Patch Set 1:

(1 comment)

I think you used the wrong variable, but is the only use of the account to login there in the test? If so, it would be much, much better to generate 16 random letters, finish the login test with the random password, and then throw it away when you're done.

Patch-set: 1
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0