Update patch set 11
Patch Set 11: Code-Review-1

(3 comments)

My main issues from merging this are still around properly sandboxing this code so that it only runs in the alpha mode of the site.

I realise coming from any other MediaWiki development the idea of a sandboxed alpha/beta/stable mode may seem strange, but it's a discipline we keep to in mobile land, and we believe it ensures our code is held up to a higher standard and avoids bugs.

I'm happy to adjust this patch myself if you prefer as I know this can be tedious and cumbersome. Let me know if you would prefer me to do this.

Patch-set: 11
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1