Update patch set 2
Patch Set 2:

Classic. I realised I'd made this mistake as soon as I shut my eyes to go to sleep last night. Somehow, because this is a non-printing character it didn't occur to me that it would make the markup invalid by being between the <li> tags.

I've re-pushed my commit with the &zwnj; inside the first <li>, and it works just as well. Thanks :)

Patch-set: 2