Update patch set 4
Patch Set 4:

Just a head's up that Jenkins-bot was not marking this patchset as fully verified because a PHP style checker was failing. I pulled the patchset down to clean up the file in question to make the checks pass, which turned this into a *much* bigger patchset, mostly with changes unrelated to the bug itself. This is kind of a bummer (I'd rather see the cleanup changes in a different patchset), but since we lost the ability to manually 'verify' patchsets, I'm not sure this will be otherwise mergable.

Patch-set: 4
Label: Code-Review=0