Hygiene: Use accept attribute in uploads file input tag
This change removes irrelevant options in the Android dialog that asks
for the app that should handle the file input. Previously Android would
display all apps, including the ones that can't be used to select images
(e.g. music related).

Tested on Android Browser 4.1 and 4.2 and it works fine, you can use
camera. We don't support uploads on Android < 4.

Change-Id: I61e056823ff3e82209de483d6d974fef3fcf672c