Add $wgMFCollapseSectionsByDefault config var to allow disabling auto section collapse
The section collapse model doesn't work very well on sites like Wiktionary,
where there usually is little or no content in section 0 and the rest of
the top-level sections are languages. Setting allows it to be disabled for
that site only, leaving sections open and readable.

Default is $wgMFCollapseSectionsByDefault = true, preserving current behavior.

Bug: 61447
Change-Id: I1b14341f74d1894fd1a0a5682d1db3d082435bbd