Update patch set 4
Patch Set 4: Code-Review-1

(8 comments)

On a side note - kaity: kaldari awjr why are we asking "Are you sure you want to create a new page on Wikipedia?" when a new page is created - seems really bizarre to have it at the end of a workflow. If I've invested a load of time to create a page why would i say no? Seems like bad design…

Patch-set: 4
Label: Code-Review=-1