Update patch set 3
Patch Set 3: Code-Review-1

I'm a bit confused. This seems to make the button on alpha smaller not wider... Also isn't the point of mw-ui-block to make buttons wide - shuldn't we be updating mediawiki ui instead?

Either the commit message can be made clearer or we're doing it wrong. We should be eradicating special cased buttons in favour of mediawiki ui classes.

Patch-set: 3
Label: Code-Review=-1