Update patch set 1
Patch Set 1:

I'm not sure this kind of things is testable. Yes, we could theoretically create a couple of files, set TS for them and test if output is the same, however this does not guarantee that result will work in practice because of complex interactions with other parts of RL. That's why core has no tests for that.

Patch-set: 1
Label: Verified=0