Update patch set 1
Patch Set 1: Code-Review-1

Caching implications

With this code users viewing cached pages will see the language section rather than the nice blue button.

You can demonstrate the caching problem with:
./scripts/cachedpage.sh cb52f3d4c0f8721f0df937250d254c0ec2966bca 'Page with languages'

Kenan may say this is not a big deal in which case I'm happy to switch over to a +2

To fix this it is pretty trivial - simply check for the language section at the top of languages.js and if it exists remove it in JavaScript and manually create the button.

// FIXME: Remove after 30 days.
var $cachedHeading = $( '#section_language' ), label;
if ( $cachedHeading.length > 0 ) {
  label = $cachedHeading.text();
  $( '<a class="mw-ui-button mw-ui-progressive button languageSelector">' ).
    text( label ).
    appendTo(
      $( '<div id="page-secondary-actions"></div>' ).insertBefore( '#mw-mf-language-section' )
    );
  $( '#mw-mf-language-section' ).remove();
}

Patch-set: 1
Label: Code-Review=-1