Ensure checkToggleView() fires at appropriate time
Resolves bug 61400. MobileContext::checkToggleView() was firing after
the first time shouldDisplayMobileView() did, rendering it partly
impotent. Now it will fire just before
shouldDisplayMobileViewInternal(), ensuring it is properly honored.

Also removes shouldDisplayMobileView() check prior to running
GetMobileUrl hook. shouldDisplayMobileView() should not get called from
getMobileUrl() (with this patch, it can cause an infinite loop), but
getMobileUrl() shoudl not care whether or not we are in mobile mode.
Bug: 61400

Change-Id: Ic94dbd73634edcd2dd35065c1e7f7a0ca03340dd