Update patch set 2
Patch Set 2: Code-Review-1

Cucumber is completely agnostic as to whether a step is preceded by Given, When or Then.  See https://github.com/cucumber/cucumber/wiki/Given-When-Then

That's not to say you didn't fix a problem with some of the other changes, but substituting And and When is not it. 

Adding Zeljko as reviewer

Patch-set: 2