Update patch set 2
Patch Set 2: Code-Review-1

Is it not better to be explicit about the page being tested? As if a test runs on a random page then the results could be inconsistent depending on the page visited.

Can we discuss this alternative approach before merging.. ?
https://gerrit.wikimedia.org/r/#/c/117371/

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1