Update patch set 1
Patch Set 1: Code-Review-1

The correct way to do attribution in the i18n file is to add a list of authors to the en array:
/** English (English)
 * @author Patrick Reilly
 */
$messages['en'] = array(

There's no need for the copyright notice. The requirement for copyright notices was removed in 1989. If we really do want to have a copyright notice, it would need to say "copyright 2011-2014 Wikimedia Foundation" as it was a work for hire (unless you wrote it as a volunteer).

Patch-set: 1
Reviewer: Kaldari <78@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1