Make sure Minerva is registered as a valid skin
Right now Minerva is only registered as a valid skin if the user is
on desktop and has Minerva activated in BetaFeatures. Minerva also
needs to be registered in mobile view so that skinStyles work
correctly. It would also be useful to be able to use useskin=minerva
regardless of whether it is turned on in BetaFeatures.

This code registers the skin no matter what and just hides it as an
option in the preferences unless Minerva is activated in BetaFeatures.

The most immediate need for this fix is allowing ooui-js themes to be
properly associated with skins (e.g. apex theme not loaded in
Minerva).

Bug: 36636
Change-Id: Ib9d52327425f8a12e12d82f1421f5cab0774c510