Update patch set 4
Patch Set 4:

If it works for you, I'm fine with it. Sorting alphabetically is one way to deal with this. Another way could be that you organise it by feature or something. The translations in other codes will be exported in the order the messages are in en.json.

Patch-set: 4
Reviewer: Siebrand <92@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0