Update patch set 1
Patch Set 1:

I should also note I am responsible for all the instances of "Main Page". To me it is more user friendly to put "Main Page" but I sometimes forget that really this is '_' - the motivation behind the change was to avoid the `customer` having to think about this. I would say without this patch we run risk of reintroducing this again - as sometimes you just forget the weird rules of MediaWiki.

"* I can see myself in the future writing a test using a string "Has space" and wondering why it is coming out "Has_space". Even worse is that the string "Has two spaces" is going to come out "Has_two spaces" because of this snippet."

That should definitely be fixed though - this is due to me not knowing Ruby  - sorry... :-S

That said this really isn't a big deal and we have much bigger problems to deal with. We should either fix up this patch so two spaces get underscores or we should forget this bug ever happened and WONTFIX it. The redirects really shouldn't matter.

Patch-set: 1