Update patch set 2
Patch Set 2: Code-Review+2

Sorry I got confused.. there are instances which should be using this code which are not. We might need to restore this later to include the needsPhoto class when this code gets more widely used but you are right - my bad - this is currently an unused variable so merging.

Patch-set: 2
Label: Code-Review=+2