Update patch set 7
Patch Set 7: Code-Review-1

(4 comments)

Still failing for me..

    Given I create a random page using the API # features/step_definitions/create_page_api_steps.rb:3
    And I am logged into the mobile website    # features/step_definitions/common_steps.rb:10
    And I am on the random page                # features/step_definitions/common_steps.rb:94
    And the page is unwatched                  # features/step_definitions/watchlist_steps.rb:1
      timed out after 5 seconds, waiting for {:text=>"Yes", :tag_name=>"button"} to become present (Watir::Wait::TimeoutError)
      ./features/step_definitions/watchlist_steps.rb:4:in `/^the page is unwatched$/'
      features/watchstar.feature:8:in `And the page is unwatched'

Note it also seems to be clicking the watch star /link/ rather than testing the JavaScript version. This is not an accurate test for the experience the majority of users will get. You might want to add a wait step before the link is clicked or use a more specific selector for the watchlist star (watch-this-article class)


watch-this-article

Patch-set: 7
Label: Code-Review=-1