Update patch set 2
Patch Set 2: Code-Review-1

Thanks for this Theo! Note this opens the editor with the toggle button to the left of the overlay heading - we'll probably want to disable this too when this is the case!

cc'ing Krinkle to see if this is the best way to test for existence of VisualEditor. Not sure if it is or not..

Patch-set: 2
Label: Code-Review=-1