Update patch set 4
Patch Set 4:

After futzing with this for some time I chatted with David Burns of Mozilla on #selenium in IRC and it boils down to a bug in the Selenium Firefox driver. 

The various Selenium parties and the W3C are working out a definition of "interactable" that will resolve this, but we won't see that any time soon. 

In the meantime, let's just run this test for Chrome. 

Also, we might want to think whether we should have in the upload page a status message AND a progress bar AND a spinner all (potentially) in the page at the same time.

Patch-set: 4
Reviewer: Cmcmahon <77@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0