Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

If you set the duration of animation to something long (2s) you can see that the right drawer is not actually animated.

This will probably require a small change of how the show() and hide() methods work in Overlay. Right now we append and detach the overlay, but we'll probably want to have it in the DOM all the time and just hide it (or translate it). If we add an element to the DOM and try to use a transition just after that, it might not work (I had this problem before with bottom drawers).

Patch-set: 1
Label: Code-Review=-1