Update patch set 2
Patch Set 2:

Sure.. So we load no JavaScript related to VisualEditor at the start of the browsing session. When the edit button is clicked a call to mw.loader.using is made which pulls down the JavaScript code...

So after this you'll be able to test whether visual editor is present .. not before. :-S

Does this make sense?

Patch-set: 2