Update patch set 1
Patch Set 1: Code-Review+1

Siebrand I'm not sure I understand the problem here. The issue is that a user has complained that the privacy text is not the same for desktop and mobile so is this not the desired result?

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1