Update patch set 3
Patch Set 3: Code-Review-1

(36 comments)

I appreciate the effort, but there are many inconsistencies (mainly when it comes to @class usage).

It seems JSDuck doesn't have a @module concept [1], but the VE team dealt with it by marking those classes that are not classes with @singleton and @override:

/**
 * @class ve.debug
 * @override ve
 * @singleton
 */

Also, it would be nice if there was a Gemfile for installing JSDuck locally in a gemset, similarly to how we have package.json for Node.js.

[1] https://github.com/senchalabs/jsduck/issues/297

Patch-set: 3
Reviewer: JGonera <542@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1