Destroy the VE target after saving
Otherwise, the target and its surface stick around, then if the user
edits again, there are two targets and two surfaces on the page,
one of them will start responding to the other's events, and lots
of weird things happen.

What should really be done is MobileViewTarget should be long-lived
and allow for multiple activations and deactivations, like ViewPageTarget.
Also, a lot of the code in VisualEditorOverlay should really be in
MobileViewTarget and should be deduplicated with code from ViewPageTarget,
right now it's a pile of hacks.

Bug: 64032
Bug: 64034
Change-Id: Ic2a4b5ded429e9b6dce8773dc60b1ff2c0307fee