Update patch set 3
Patch Set 3: Code-Review-1

I worry we are making our code a mess for the case of ProofRead page on wikisource. As discussed with Max it's important we fix this asap, but I think we should simply disable SpecialMobileDiff on Wikisource.

Give me until this afternoon to propose an alternative fix for this problem.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1