Update patch set 3
Patch Set 3:

Sorry to be unclear in our discussion, but I wasn't saying you shouldn't do a check for anonymous users, but that you should call $user->getEditToken() and $user->matchEditToken() even if the user is anonymous.

Also, adding "mobile" as a salt wouldn't hurt for logged in users.

Patch-set: 3