LanguageOverlay: Only display title if available
Some pages (for example, the Main Page on the English Wikipedia)
do not include page titles for the page in other languages.

Mustache {{#x}}{{/x}} shows the content if x is not directly
false, but the API returns an empty string, which evaluates
as true. We set item.title with the title or false, which both
handles the issue and results in more descriptive variables.

Change-Id: I05949e4ea74094cb24af612456dd1a8ed74861f8