Update patch set 7
Patch Set 7: Code-Review+1

The styling appears to be different when you compare beta to stable (due to the typography changes that has currently not been pushed to stable)

I'm not sure if that is a deal breaker.

Only putting as a +1 as all the dependencies need to be merged first.

Patch-set: 7
Label: Code-Review=+1