Replace parentheses with pipe in the language list
Currently the article name is put in parentheses in the language list.
This is problematic for several reasons:
1. Different languages use parentheses differently - different meaning,
   different spacing, different unicode characters, etc.
2. In RTL wikis this breaks in a major way. For an example see:
   https://he.m.wikipedia.org/wiki/ASCII#/languages
3. Many article names have parentheses in their names, such as
   [[Georgia (country)]].

This commit resolves most of these issues in a simple way by
removing the parentheses and adding a pipe character to separate
the language name and the article title. This should be generic
enough to show in all languages.

This doesn't resolve all RTL issues in the language list,
but it does remove the most notable one.

Change-Id: I712b629da95400432ab3f02d7efbad7dd6c38556