Update patch set 1
Patch Set 1: Code-Review-1

Can you cut a bug for this and tag it design as I'm not sure about this. It doesn't look very good on a tablet screen for one and it also creates an inconsistency with the other lists we have e.g. nearby, watchlist and pages. Adding a small amount of CSS for one specific use case seems like a bad code smell so I'd like us to be a bit more cautious here.

Would it not be better to add dir attributes to list items?

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1