Make sure mobile VisualEditor is only loaded when appropriate
On wikis that are VE opt-in, show the wikitext editor by default.
On wikis that are VE opt-out, show the VisualEditor by default.

Make sure we are only showing for appropriate namespaces and
not for pages that are the output of the Page Translation feature.

Make sure the switcher bar displays above the VE toolbar
elements.

Update browser tests appropriately.

Bug: 64681
Change-Id: Id801ae13206e8d927d06997b146c201f869758f3