Update patch set 16
Patch Set 16:

I'm not sure how this can be improved for non-JavaScript phones.... Yes the data URIS will be in the CSS file but the rules will never be run. We could distangle the icons so they only load with JS but this seems a bit overkill.

I'm not sure what would break on older devices.

It's worth noting these rules have been like this for some time and I'm not doing anything new here (previously we overrode the background images to be smaller resolution ones and now we are simply doing the same but specifying background none)

Patch-set: 16
Label: Verified=0