QA: Give VE more time to load in browser tests
When switching from the wikitext editor to VisualEditor we give the
default 5 seconds of loading time for it to load. During this time
not only does it need to load all its JavaScript but it also needs
to re-initialise. Are you crazy Cucumber? 5 seconds to do all this magic?!

Give it 20 seconds - much more reasonable and let our browsers test pass
again.

Change-Id: I8896e834a1fb074fab4f40ac22019864567cb6d0