Update patch set 2
Patch Set 2:

I don't think this is a good idea - the host is a parameter for this function, and it might change between calls. The only reasonable optimization would be to store host in a static var, and compare it with the previous value, and if they match, avoid testing. This might be an overkill and won't have any measurable perf impact.

Patch-set: 2
Reviewer: Yurik <571@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0