Update patch set 1
Patch Set 1: Code-Review-1

Two things.

First, you know, I am a bit of a goof. I forgot we do the minifying automatically. I will update the Storer.js repo with a non-minified version of Storer-IEless.js. Second, if you are using the IEless version, please keep "ieless" in the file name so that we don't mistake it for the full version.

Patch-set: 1
Label: Code-Review=-1