Prevent horizontal scrollbar after menu closing animation
If you close the menu with animations enabled, you can scroll
horizontally even thought there is not content to the right of the page.
This is caused by removing overflow: hidden while the transition is
still in progress.

The page gets wider when we open the nav because we translate
nav is open if it wasn't for overflow: hidden in .navigation-enabled
overflow: hidden is no longer applied and the horizontal scrollbar
appears just when the transition starts. Unfortunately, it stays there
because, I guess, Webkit doesn't recalculate the size of

overflow-x: hidden should be always safe since we force wrapping text.

Also, fix comment near overflow: hidden rule, it prevents both
horizontal and vertical scrolling.

Change-Id: Ia8d9e4228cc0109d7234c0612270081322786014