Update patch set 1
Patch Set 1:

Hm.. I don't think having it report fail/pass based on how much has changed is useful. Font rendering can indeed be variable, but shouldn't change when in the same environment (e.g. when the screenshot is made by selenium/web driver). If it does we can indeed make it tolerate the slightest amount to accommodate for that, but I fear it'll mask tiny changes, the kind you wouldn't normally see straight away, but a tool like this would immediately catch (e.g. a 1px border no longer being there because of a typo, or it changing colour).

It wouldn't report a failure, that's only asking for extra maintenance burden (what if you did want to make a large change, do you need to override Jenkins -2? or invent some silly code to detect in the commit message?).

When I used pdiff in another project, the power was that if it reported any difference other than an exact px match, you'd check it out (on your own time, between submitting to Jenkins and getting reviewed) and make sure it looks as intended. Helped catch lots of small things you can't easily catch otherwise. A nice catch all. I think having it fail on certain ratios will create a habit of not looking at it when it's low enough. And it only works if you use it for each change.

Patch-set: 1