Fix watchlist query to work with PostgreSQL
MySQL allows SQL to select a column which is not included in the
GROUP BY, and accomodates this by returning an arbitrary value
for that column from each group.  PostgreSQL does not allow this
situation.  Fix this by explicitly listing the columns to be
selected rather than using '.*' notation.

There is one other place in this extension where '.*' is used,
but that is not in a GROUP BY query and so should not cause
this type of problem.

Bug: 46745
Change-Id: I56caca2666be5eb94dac4a3d316a5bb140cb7b48