Update patch set 1
Patch Set 1:

Aside from my comments the one big problem with this patchset is a change in behaviour. Previously going to Special:Uploads whilst anonymous would prompt you to login. This was useful. Now it says User "" is not registered. This is not useful and a change in behaviour.

Patch-set: 1