Update patch set 1
Patch Set 1:

> Something smells fishy here - why does this happen

Probably due to a Webkit's bug or ambiguity in CSS transitions specs.

> and why are we preventing horizontal scrolling all the time?

Why not? It should never happen. If something is wider than the viewport, the whole site looks messed up. And I think we got rid of any instances of that (max-width on images, wrapping very long words).

> I'm not sure adding overflow-x maps directly to the problem. Is there something funny with border-box going on? (-1ing purely to get your attention)

No, not border box. The page gets wider when we open the nav because we translate #mw-mf-page-center to the right. The scrollbar would be shown when the nav is open if it wasn't for overflow: hidden in .navigation-enabled #mw-mf-viewport. However, as soon as we remove .navigation-enabled, overflow: hidden is no longer applied and the horizontal scrollbar appears just when the transition starts. Unfortunately, it stays there because, I guess, Webkit doesn't recalculate the size of #mw-mf-viewport's content after the transition finishes.

Patch-set: 1
Reviewer: JGonera <542@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0