Update patch set 3
Patch Set 3:

External code should be tested, but it should be tested in the external repo.  Those tests can then be run before a new version is imported into the repo (since it's an external library, people should not be changing it here; if they do it becomes a fork/tracking branch, not external code)

jQuery is critical functionality for us, and it is tested (https://github.com/jquery/jquery/tree/master/test), but it shouldn't be tested in every repo that uses jQuery.

Patch-set: 3
Label: Verified=0